requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-environ==0.11.2
mysqlclient==2.2.4
scikit-learn>=1.5.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0
Pillow>=10.0.0
